Split library into different files to simplify maintenance
